Moved icons to participants folder.
