+++ src/components/studies/participants/PageSelector.tsx
@@ -1,8 +1,8 @@
 import React from 'react'
-import NextPageIcon from '../../../assets/ParticipantManagerPageSelector/next_page_icon.svg'
-import PreviousPageIcon from '../../../assets/ParticipantManagerPageSelector/previous_page_icon.svg'
-import BackToBeginningIcon from '../../../assets/ParticipantManagerPageSelector/back_to_beginning_icon.svg'
-import ForwardToEndIcon from '../../../assets/ParticipantManagerPageSelector/forward_to_end_icon.svg'
+import NextPageIcon from '../../../assets/participants/paging/next_page_icon.svg'
+import PreviousPageIcon from '../../../assets/participants/paging/previous_page_icon.svg'
+import BackToBeginningIcon from '../../../assets/participants/paging/back_to_beginning_icon.svg'
+import ForwardToEndIcon from '../../../assets/participants/paging/forward_to_end_icon.svg'
 import PageBox from './PageBox'
 import {makeStyles} from '@material-ui/core/styles'
 import {Button} from '@material-ui/core'
