+++ src/components/studies/participants/ParticipantManagerPlaceholder.tsx
@@ -1,9 +1,9 @@
 import React from 'react'
 import {makeStyles, Box} from '@material-ui/core'
-import RocketIcon from '../../../assets/participant-manager/participant_manager_rocket_icon.svg'
-import LiveCard from '../../../assets/participant-manager/participant_manager_live_card_draft_icon.svg'
-import DraftCard from '../../../assets/participant-manager/participant_manager_study_card_draft_icon.svg'
-import RightArrow from '../../../assets/participant-manager/right_arrow.svg'
+import RocketIcon from '../../../assets/participants/participant_manager_rocket_icon.svg'
+import LiveCard from '../../../assets/participants/participant_manager_live_card_draft_icon.svg'
+import DraftCard from '../../../assets/participants/participant_manager_study_card_draft_icon.svg'
+import RightArrow from '../../../assets/participants/right_arrow.svg'
 import {latoFont} from '../../../style/theme'
 import clsx from 'clsx'
